fix: unsigned short -> uint16_t

+++ src/kernel/include/string.h
@@ -4,8 +4,8 @@
 #include "system.h"
 
 extern void*	memcpy(void* to, const void* from, size_t count);
-extern uint8_t* memset(uint8_t* dest, uint8_t val, size_t count);
-extern uint8_t* memsetw(uint8_t* dest, uint8_t val, size_t count);
+extern uint8_t* memset(uint8_t* dest, uint16_t val, size_t count);
+extern uint8_t* memsetw(uint8_t* dest, uint16_t val, size_t count);
 extern size_t	strlen(const uint8_t* str);
 
 #endif
