+++ src/kernel/string/memsetw.c
@@ -1,6 +1,6 @@
 #include <string.h>
 
-uint8_t* memsetw(uint8_t* dest, uint8_t val, size_t count)
+uint8_t* memsetw(uint8_t* dest, uint16_t val, size_t count)
 {
 	/* Same as above, but this time, we're working with a 16-bit
 	 *  'val' and dest pointer. Your code can be an exact copy of
